Add Aspire.Hosting.IIS
This adds resources to manage IISExpress. This will work in Visual Studio with old-style csproj files and on the command-line with SDK style csproj files. Debugging is not supported out of the box (similar to python and node apps), but debuggers can be attached to it.

Heavily inspired from https://github.com/CZEMacLeod/C3D.Extensions.Aspire but simplified for common use cases.